# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecurityPreflight API",
    "version": "1.0.0",
    "description": "Machine-readable OpenAPI specification for this service.",
    "license": {
      "name": "Proprietary",
      "identifier": "LicenseRef-Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8781",
      "description": "Local Docker Desktop API server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid health check request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness check",
        "operationId": "getReadiness",
        "tags": ["System"],
        "security": [],
        "responses": {
          "200": {
            "description": "Service is ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid readiness check request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service is not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "summary": "List registered projects",
        "operationId": "listProjects",
        "tags": ["Projects"],
        "security": [],
        "responses": {
          "200": {
            "description": "Registered projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid projects request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scan-profiles": {
      "get": {
        "summary": "List built-in scan profiles",
        "operationId": "listScanProfiles",
        "tags": ["Scans"],
        "security": [],
        "responses": {
          "200": {
            "description": "Available scan profiles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanProfileListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid scan profile request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/toolchain/doctor": {
      "get": {
        "summary": "Check local scanner and runtime tool availability",
        "operationId": "getToolchainDoctor",
        "tags": ["Toolchain"],
        "security": [],
        "responses": {
          "200": {
            "description": "Toolchain diagnostic result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolchainDoctorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid toolchain doctor request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "required": ["status", "service", "version", "timestamp"],
        "properties": {
          "status": {
            "type": "string",
            "examples": ["ok"]
          },
          "service": {
            "type": "string",
            "examples": ["security-preflight"]
          },
          "version": {
            "type": "string",
            "examples": ["1.0.0"]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "path",
          "repositoryUrl",
          "defaultBranch",
          "technologyStack",
          "dataClassification",
          "owner",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "repositoryUrl": {
            "type": ["string", "null"]
          },
          "defaultBranch": {
            "type": ["string", "null"]
          },
          "technologyStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dataClassification": {
            "type": "string",
            "enum": ["public", "internal", "confidential", "sensitive", "health-data"]
          },
          "owner": {
            "type": ["string", "null"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectListResponse": {
        "type": "object",
        "required": ["data", "meta"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "meta": {
            "type": "object",
            "required": ["total"],
            "properties": {
              "total": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        }
      },
      "ScanProfile": {
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "checks",
          "failThreshold",
          "allowActiveDast",
          "allowProductionTargets",
          "timeoutSeconds"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "checks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "failThreshold": {
            "type": "string",
            "enum": ["critical", "high", "medium", "low"]
          },
          "allowActiveDast": {
            "type": "boolean"
          },
          "allowProductionTargets": {
            "type": "boolean"
          },
          "timeoutSeconds": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ScanProfileListResponse": {
        "type": "object",
        "required": ["data"],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScanProfile"
            }
          }
        }
      },
      "ToolchainDoctorResponse": {
        "type": "object",
        "required": ["checkedAt", "tools", "summary"],
        "properties": {
          "checkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToolCheck"
            }
          },
          "summary": {
            "type": "object",
            "required": ["available", "missing", "error"],
            "properties": {
              "available": {
                "type": "integer",
                "minimum": 0
              },
              "missing": {
                "type": "integer",
                "minimum": 0
              },
              "error": {
                "type": "integer",
                "minimum": 0
              }
            }
          }
        }
      },
      "ToolCheck": {
        "type": "object",
        "required": ["name", "command", "status", "version", "message"],
        "properties": {
          "name": {
            "type": "string"
          },
          "command": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["available", "missing", "error"]
          },
          "version": {
            "type": ["string", "null"]
          },
          "message": {
            "type": ["string", "null"]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string",
                "examples": ["VALIDATION_ERROR"]
              },
              "message": {
                "type": "string",
                "examples": ["Invalid request data."]
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              },
              "requestId": {
                "type": "string",
                "examples": ["req_abc123"]
              }
            }
          }
        }
      }
    },
    "securitySchemes": {}
  },
  "tags": [
    {
      "name": "System",
      "description": "System endpoints used for health, readiness and diagnostics."
    },
    {
      "name": "Projects",
      "description": "Local project registration and metadata."
    },
    {
      "name": "Scans",
      "description": "Scan profiles and scan run orchestration."
    },
    {
      "name": "Toolchain",
      "description": "Local scanner and runtime diagnostics."
    }
  ]
}